+++ src/main/java/org/jamesframework/core/subset/neigh/SingleSwapNeighbourhood.java
@@ -22,9 +22,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
 import org.jamesframework.core.subset.SubsetSolution;
+import org.jamesframework.core.util.Randomization;
 import org.jamesframework.core.util.SetUtilities;
 
 /**
@@ -78,8 +78,8 @@ public class SingleSwapNeighbourhood extends SubsetNeighbourhood {
             // impossible to perform a swap
             return null;
         }
-        // use thread local random for better concurrent performance
-        Random rg = ThreadLocalRandom.current();
+        // retrieve random generator
+        Random rg = Randomization.getRandom();
         // select random ID to remove from selection
         int del = SetUtilities.getRandomElement(removeCandidates, rg);
         // select random ID to add to selection