+++ src/main/java/org/jamesframework/core/subset/neigh/adv/MultiDeletionNeighbourhood.java
@@ -22,12 +22,12 @@ import org.jamesframework.core.subset.neigh.moves.GeneralSubsetMove;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.IntStream;
 import org.jamesframework.core.subset.SubsetSolution;
 import org.jamesframework.core.subset.neigh.SingleDeletionNeighbourhood;
 import org.jamesframework.core.subset.neigh.moves.SubsetMove;
 import org.jamesframework.core.subset.neigh.SubsetNeighbourhood;
+import org.jamesframework.core.util.Randomization;
 import org.jamesframework.core.util.SetUtilities;
 import org.jamesframework.core.util.SubsetIterator;
 
@@ -174,8 +174,8 @@ public class MultiDeletionNeighbourhood extends SubsetNeighbourhood {
         if(curMaxDel == 0){
             return null;
         }
-        // use thread local random for better concurrent performance
-        Random rg = ThreadLocalRandom.current();
+        // retrieve random generator
+        Random rg = Randomization.getRandom();
         // pick number of deletions (in [1, curMaxDel])
         int numDel = rg.nextInt(curMaxDel) + 1;
         // pick random IDs to remove from selection